Rename FluidSynth source files to avoid include name clash.
On case insensitive filesystems the FluidSynth.h of the module collides
with the fluidsynth.h of FluidSynth and breaks the build.
