+++ Source/ModuleFactory.cpp
@@ -274,7 +274,7 @@
 #include "LatencyCalculator.h"
 
 #ifdef BESPOKE_FLUIDSYNTH
-#include "FluidSynth.h"
+#include "FluidSynthModule.h"
 #endif
 
 #include <juce_core/juce_core.h>
